refactor(test): splits get-dependencies spec
- there was a lot of things going on in get-dependencies - splitting it will make it easier to find things
- eslint complained about a too high number of top level tests because of this
